src/ai/requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
langgraph
motor
python-dotenv
pydantic
google-generativeai
python-multipart
